pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "metis"
version = "0.1.0"
description = "Semantic knowledge infrastructure for building graph databases with ArangoDB and embedding models"
authors = ["Metis Contributors"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://github.com/metis-ai/metis"
repository = "https://github.com/metis-ai/metis"
keywords = ["semantic-search", "embeddings", "graph-database", "arangodb", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
]
packages = [{ include = "metis" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"  # triton requires <3.14
httpx = "^0.27.0"
h2 = "^4.1.0"
numpy = "^1.26.0"
sentence-transformers = "^5.1.1"
torch = "^2.0.0"
tree-sitter = "^0.25.0"
tree-sitter-languages = "^1.10.0"
pyyaml = "^6.0"
grpcio = "^1.64.0"
docling = {version = "^2.54.0", optional = true}
torch-geometric = {version = "^2.6.0", optional = true}
peft = "^0.17.1"
torchvision = "^0.23.0"
arxiv = "^2.1.0"
PyGithub = "^2.1.1"
requests = "^2.31.0"
plotly = "^5.18.0"
umap-learn = "^0.5.5"
waybackpy = "^3.0.6"
kaleido = "0.2.1"
pymupdf = "^1.26.4"

[tool.poetry.extras]
pdf = ["docling"]
gnn = ["torch-geometric"]
all = ["docling", "torch-geometric"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
ruff = "^0.6.3"
mypy = "^1.11.0"
types-pyyaml = "^6.0.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "W",   # pycodestyle warnings
  "I",   # isort
]

[tool.pytest.ini_options]
addopts = "-q --cov=metis --cov-report=term-missing"
pythonpath = ["."]
testpaths = ["tests"]
